fixup! [#691] changed searchbar in menubar to mobile searchbar
